The location of BanUser function is changed from Authentication to User.
